backend/pyproject.toml
[tool.poetry]
name = "agro-ai-backend"
version = "1.0.0"
description = "Agro AI Platform Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
pillow = "^10.0.1"
torch = "^2.1.0"
torchvision = "^0.16.0"
transformers = "^4.35.0"
scikit-learn = "^1.3.2"
numpy = "^1.24.3"
opencv-python = "^4.8.1.78"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
joblib = "^1.3.2"
pandas = "^2.1.3"
aiofiles = "^23.2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"